refactor(DEX-4083): header was refactored deleting unnecessary files keeping only one header and styling the header to fit the design
