+++ src/assets/MenuIcon.jsx
@@ -1,0 +1,15 @@
+const MenuIcon = (props) => (
+  <svg
+    width="12px"
+    height="12px"
+    viewBox="0 0 16 16"
+    version="1.0"
+    {...props}
+  >
+    <rect fill="white" x="1" y="2.5" width="10" height="1" />
+    <rect fill="white" x="1" y="5.5" width="10" height="1" />
+    <rect fill="white" x="1" y="8.5" width="10" height="1" />
+  </svg>
+);
+
+export default MenuIcon;
